изчисляване на средна доставна цена

+++ price/ProductCosts.class.php
@@ -168,6 +168,7 @@ class price_ProductCosts extends core_Manager
         foreach ($tmpArr as $index => $r) {
             $pId = acc_Items::fetchField($index, 'objectId');
             $amount = (!$r->quantity) ? 0 : round($r->amount / $r->quantity, 5);
+            $r->quantity = (!$r->quantity) ? 0 : $r->quantity;
             $res[$pId] = (object)array('amount' => $amount, 'quantity' => $r->quantity);
         }
        
@@ -458,6 +459,9 @@ class price_ProductCosts extends core_Manager
         // Намираме цените по последна рецепта
         $res['bom'] = $this->getLastBomCosts($productKeys);
         
+        // Намиране на средните цени
+        $res['average'] = $this->getAveragePrices($productKeys, $res['accCost']);
+        
         // Тук ще събираме готовите записи
         $nRes = array();
         $today = dt::now();
@@ -503,6 +507,14 @@ class price_ProductCosts extends core_Manager
                 $nRes[] = $obj;
             }
             
+            if (isset($res['average'][$productId])) {
+                $obj = clone $bObject;
+                $obj->type = 'average';
+                $obj->price = $res['average'][$productId]->price;
+                $obj->quantity = $res['average'][$productId]->quantity;
+                $nRes[] = $obj;
+            }
+            
             if (isset($res['bom'][$productId])) {
                 $obj = clone $bObject;
                 $obj->type = 'bom';
@@ -528,7 +540,13 @@ class price_ProductCosts extends core_Manager
         $this->saveArray($synced['update']);
         
         if (count($synced['delete'])) {
+            $query = self::getQuery();
+            $query->in('id', $synced['delete']);
+            $query->show('type');
+            $arr = $query->fetchAll();
+           
             foreach ($synced['delete'] as $id) {
+                if($arr[$id]->type == 'average') continue;
                 $this->delete($id);
             }
         }
@@ -547,9 +565,81 @@ class price_ProductCosts extends core_Manager
     {
         expect($productId);
         expect(in_array($priceType, array('accCost', 'lastDelivery', 'activeDelivery', 'lastQuote', 'bom',)));
-        
         $price = static::fetchField("#productId = {$productId} AND #type = '{$priceType}'", 'price');
         
         return $price;
+    }
+    
+    
+    /**
+     * Средната доставна цена
+     * 
+     * @param array $productKeys
+     * @param array $accCosts
+     * @return stdClass $res
+     */
+    public function getAveragePrices($productKeys, $accCosts)
+    {
+        // Кои пера са участвали в дебитирането на склада в последните 3 месеца
+        $beforeDate = dt::addMonths(-3);
+        $jQuery = acc_JournalDetails::getQuery();
+        $jQuery->EXT('valior', 'acc_Journal', 'externalKey=journalId');
+        $storeAccId = acc_Accounts::getRecBySystemId('321')->id;
+        $jQuery->where("#debitAccId = {$storeAccId} AND #valior >= '{$beforeDate}'");
+        $jQuery->show('debitItem2');
+        $jQuery->groupBy('debitItem2');
+        $itemsWithMovement = arr::extractValuesFromArray($jQuery->fetchAll(), 'debitItem2');
+        
+        // Кои ид-та на артикули, съответстват на тези пера
+        $iQuery = acc_Items::getQuery();
+        $iQuery->where("#state = 'active' AND #classId=" . cat_Products::getClassId());
+        $iQuery->in("id", $itemsWithMovement);
+        $iQuery->in("objectId", $productKeys);
+        $iQuery->show('id,objectId');
+        $productArr = arr::extractValuesFromArray($iQuery->fetchAll(), 'objectId');
+        
+        // Извличане на данните за покупки за тези артикули
+        $res = array();
+        $purQuery = purchase_PurchasesData::getQuery();
+        $purQuery->in('productId', $productArr);
+        $purQuery->where("#state != 'rejected'");
+        $purQuery->show('quantity,price,productId');
+        $purQuery->orderBy('valior', "DESC");
+        $all = $purQuery->fetchAll();
+        
+        // Нормализираме записите
+        foreach ($productKeys as $productId) {
+            
+            // Всички покупки на търсения артикул
+            $accObject = $accCosts[$productId];
+            $foundIn = array_filter($all, function ($a) use ($productId){return $a->productId == $productId && $a->quantity >= 0;});
+            
+            // Ако има к-ва в складовата наличност
+            $averageAmount = 0;
+            if(!empty($accObject->quantity)){
+                $neededQuantity = $accObject->quantity;
+                $sum = 0;
+                foreach ($foundIn as $delData){
+                    $neededQuantity -= $delData->quantity;
+                    $sum += $delData->quantity * $delData->price;
+                    
+                    if($neededQuantity <= 0) break;
+                }
+               
+                $averageAmount = round($sum / $accObject->quantity, 4);
+                $averageAmount = core_Math::roundNumber($averageAmount);
+            } else {
+                if(count($foundIn)){
+                    $foundIn = $foundIn[key($foundIn)];
+                    $averageAmount = $foundIn->price;
+                }
+            }
+            
+            $res[$productId] = (object) array('price' => $averageAmount,
+                                              'quantity' => $accObject->quantity,
+            );
+        }
+        
+        return $res;
     }
 }
